requirements.txt
torch
torchvision
numpy
Pillow
einops
timm
ml-collections
scipy
